+++ R/simulation-spades.R
@@ -1408,15 +1408,13 @@ recoverModePre <- function(sim, rmo = NULL, allObjNames = NULL, recoverMode) {
     rmo$recoverableObjs <- rmo$recoverableObjs[seq_len(recoverMode - 1)]
   }
 
-
   if (length(sim@events) > 0) {
-
     objsInSimListAndModule <- ls(sim) %in% allObjNames[[sim@events[[1]][["moduleName"]]  ]]
     # This makes a copy of the objects that are needed, and adds them to the list of rmo$recoverableObjs
     mess <- capture.output(type = "message",
                            rmo$recoverableObjs <- append(list(if (any(objsInSimListAndModule)) {
                              Copy(mget(ls(sim)[objsInSimListAndModule], envir = sim@.xData),
-                                  filebackedDir = tempdir2("._rmo"))
+                                  filebackedDir = file.path(getOption("spades.scratchPath"), "._rmo"))
                            } else {
                              list()
                            }), rmo$recoverableObjs)
@@ -1424,7 +1422,6 @@ recoverModePre <- function(sim, rmo = NULL, allObjNames = NULL, recoverMode) {
     mess <- grep("Hardlinked version", mess, invert = TRUE)
     if (length(mess) > 0)
       lapply(mess, message)
-
   }
   endTime <- Sys.time()
   rmo$preEvents <- sim@events