+++ tests/testthat/test-paths.R
@@ -18,8 +18,9 @@ test_that("paths file does not work correctly", {
                  inputPath = getPaths()[["inputPath"]],
                  modulePath = paths$modulePath,
                  outputPath = getPaths()[["outputPath"]],
-                 rasterPath = getPaths()[["rasterPath"]]
-                 ), normPath)
+                 rasterPath = getPaths()[["rasterPath"]],
+                 scratchPath = getPaths()[["scratchPath"]]
+                ), normPath)
               )
 
   # test for non consecutive order, but named
@@ -31,13 +32,15 @@ test_that("paths file does not work correctly", {
                            inputPath = getPaths()[["inputPath"]],
                            modulePath = paths$modulePath,
                            outputPath = path.expand(paths$outputPath),
-                           rasterPath = getPaths()[["rasterPath"]]
+                           rasterPath = getPaths()[["rasterPath"]],
+                           scratchPath = getPaths()[["scratchPath"]]
                ), normPath))
 
   # test for all unnamed
   paths <- list(tmpdir,
                 tmpdir,
                 system.file("sampleModules", package = "SpaDES.core"),
+                tmpdir,
                 tmpdir,
                 tmpdir)
   mySim <- simInit(times, params, modules, objects = list(), paths)
@@ -46,7 +49,8 @@ test_that("paths file does not work correctly", {
                            inputPath = paths[[2]],
                            modulePath = paths[[3]],
                            outputPath = paths[[4]],
-                           rasterPath = paths[[5]]
+                           rasterPath = paths[[5]],
+                           scratchPath = paths[[6]]
                ), normPath))
 
   # test for all named, non consecutive, using accessors
@@ -54,29 +58,29 @@ test_that("paths file does not work correctly", {
                 modulePath = system.file("sampleModules", package = "SpaDES.core"),
                 outputPath = tmpdir,
                 inputPath = tmpdir,
-                rasterPath = tmpdir)
+                rasterPath = file.path(tmpdir, "raster"), ## subdir of scratchPath
+                scratchPath = tmpdir)
   mySim <- simInit(times, params, modules, objects = list(), paths)
   expect_equal(lapply(paths(mySim), normPath),
                lapply(list(cachePath = cachePath(mySim),
                            inputPath = inputPath(mySim),
                            modulePath = modulePath(mySim),
                            outputPath = outputPath(mySim),
-                           rasterPath = rasterPath(mySim)
+                           rasterPath = rasterPath(mySim),
+                           scratchPath = scratchPath(mySim)
                ), normPath))
 
   # missing paths
   oldPaths <- getPaths()
   do.call(setPaths, paths)
-  on.exit({do.call(setPaths, paths)}, add = TRUE)
+  on.exit({do.call(setPaths, oldPaths)}, add = TRUE)
   mySim1 <- simInit(times, params, modules, objects = list())
-  expect_equal(lapply(paths(mySim), normPath),
-               lapply(paths(mySim1), normPath))
+  expect_equal(lapply(paths(mySim), normPath), lapply(paths(mySim1), normPath))
 
   # setting paths via setPaths works for changing the paths in the simInit & therefore simList
   setPaths(cachePath = oldPaths$cachePath)
   mySim2 <- simInit(times, params, modules, objects = list())
-  expect_false(identical(lapply(paths(mySim), normPath),
-               lapply(paths(mySim2), normPath)))
+  expect_false(identical(lapply(paths(mySim), normPath), lapply(paths(mySim2), normPath)))
 
   cachePath(mySim) <- tmpdir
   expect_equal(cachePath(mySim), tmpdir)
@@ -92,4 +96,7 @@ test_that("paths file does not work correctly", {
 
   rasterPath(mySim) <- tmpdir
   expect_equal(rasterPath(mySim), tmpdir)
+
+  scratchPath(mySim) <- tmpdir
+  expect_equal(scratchPath(mySim), tmpdir)
 })