+++ NEWS.md
@@ -9,6 +9,9 @@ version 1.0.10.9000
   but should restore the original value `on.exit`.
 * `saveSimList()` and `loadSimList()` accept `.qs` or `.rds` files
 * `spades` and `simInit` now force UTF-8 encoding; this is reset `on.exit`. If a module needs a different character encoding, then it can be set within the module code.
+* new option `spades.scratchPath`, to be used for e.g., temporary raster files and temparary SpaDES recovery mode objects.
+  The default temporary `rasterTmpDir` has changed to be a subdirectory of `scratchPath`.
+  **`rasterPath` will be deprecated in a future release.**
 
 ## dependency changes
 * none
@@ -16,6 +19,7 @@ version 1.0.10.9000
 ## bug fixes
 * improved handling of `data.table` objects using `loadSimList()`
 * caching of `.inputObjects` did not correctly capture objects that were assigned to `mod$xxx`. This is now fixed.
+* fix recovery mode bug: use scratch directory specified by the user via `options(spades.scratchPath)` (see above).
 
 version 1.0.10
 ==============